embedding source directory
