+++ source/static/styles.css
@@ -1,0 +1,396 @@
+* {
+	padding: 0;
+	margin: 0;
+	box-sizing: border-box;
+  }
+  
+  html,
+  body {
+	height: 100%;
+	width: 100%;
+	font-family: Verdana, Arial, Tahoma, sans-serif;
+	font-size: 1rem;
+  }
+  
+  main {
+	height: 100%;
+	width: 100%;
+  }
+  
+  .btn {
+	border: none;
+	border-radius: 0.25rem;
+	padding: 0.5rem 1rem;
+	cursor: pointer;
+	background: var(--tw-green-500);
+	color: var(--tw-slate-100);
+	font-weight: bold;
+  }
+  
+  .btn-ghost {
+	padding: 0;
+	border: none;
+	cursor: pointer;
+	color: var(--tw-indigo-500);
+	background: none;
+	font-weight: bold;
+	font-size: 1rem;
+  }
+  
+  .container {
+	margin: 0 auto;
+	padding: 0 1rem;
+	max-width: 1440px;
+  }
+  
+  .nav {
+	padding: 1.5rem 0;
+	position: fixed;
+	width: 100%;
+	box-shadow: 0 25px 50px -12px rgb(0 0 0 / 0.25);
+  }
+  
+  .nav__content {
+	display: flex;
+	justify-content: space-between;
+	align-items: center;
+  }
+  
+  .nav__brand {
+	display: flex;
+	align-items: center;
+	color: var(--tw-slate-100);
+	font-size: 1.25rem;
+	font-weight: bold;
+	text-decoration: none;
+	text-transform: uppercase;
+  }
+  
+  .nav__brand>svg {
+	height: 1.8rem;
+	width: 1.8rem;
+  }
+  
+  .nav__list {
+	width: 325px;
+	display: flex;
+	list-style: none;
+	justify-content: flex-end;
+  }
+  
+  .nav__item {
+	margin-right: 1rem;
+	min-width: fit-content;
+  }
+  
+  .nav__item:last-child {
+	margin: 0;
+  }
+  
+  .nav__link {
+	cursor: pointer;
+	color: var(--tw-slate-100);
+	text-decoration: none;
+	text-transform: uppercase;
+	background: none;
+	border: none;
+	font-size: 1rem;
+  }
+  
+  .nav__link:hover {
+	border-bottom: 2px solid var(--tw-slate-100);
+  }
+  
+  .hero {
+	padding: 1.5rem;
+	height: 100%;
+	width: 100%;
+	display: flex;
+	flex-direction: column;
+	justify-content: center;
+	align-items: center;
+	color: var(--tw-slate-100);
+	background: linear-gradient(
+	  to right,
+	  var(--tw-indigo-700),
+	  var(--tw-blue-600),
+	  var(--tw-indigo-700));
+  }
+  
+  .hero__title {
+	padding-bottom: 2rem;
+	font-size: 1.75rem;
+  }
+  
+  .hero__intro {
+	padding-bottom: 2rem;
+	font-size: 1.2rem;
+  }
+  
+  .waitlist-form {
+	display: flex;
+	flex-direction: column;
+	min-width: 22rem;
+  }
+  
+  .waitlist-form__group {
+	padding-top: 1rem;
+  }
+  
+  .waitlist-form__label {
+  }
+  
+  .waitlist-form__input {
+	padding: 0.5rem;
+	width: 100%;
+	margin-bottom: 1rem;
+	border: none;
+	border-radius: 0.25rem;
+  }
+  
+  .waitlist-form__message {
+	height: 2.25rem;
+	padding-top: 1rem;
+	padding-bottom: 0;
+	text-align: center;
+	font-size: 1rem;
+  }
+  
+  .waitlist-form__message-error {
+	color: var(--tw-slate-100);
+  }
+  
+  .waitlist-form__btn {
+	width: 100%;
+  }
+  
+  .waitlist-form__message,
+  .waitlist-form__message-error {
+	text-align: center;
+	margin-top: 0.5rem;
+  }
+  
+  .auth-form__wrapper {
+	padding: 0 1rem;
+	width: 100%;
+	height: 100%;
+	display: flex;
+	flex-direction: column;
+	justify-content: center;
+	align-items: center;
+  }
+  
+  .auth-form {
+	padding: 2.75rem;
+	display: flex;
+	flex-direction: column;
+	border: solid 2px var(--tw-slate-900);
+	border-radius: 0.25rem;
+  }
+  
+  .auth-form__title {
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	color: var(--tw-indigo-600);
+	font-size: 1.75rem;
+	font-weight: bold;
+	text-decoration: none;
+	text-transform: uppercase;
+	margin-bottom: 1rem;
+  }
+  
+  .auth-form__title>svg {
+	height: 2.8rem;
+	width: 2.8rem;
+  }
+  
+  .auth-form__group {
+	padding-top: 0.5rem;
+  }
+  
+  .auth-form__label {
+	margin-bottom: 0.25rem;
+	color: var(--tw-slate-900);
+  }
+  
+  .auth-form__input {
+	font-size: 1.1rem;
+	padding: 0.5rem;
+	width: 100%;
+	margin: 0.5rem 0;
+	border: solid 1px var(--tw-slate-900);
+	border-radius: 0.25rem;
+  }
+  
+  .auth-form__message {
+	height: 2.25rem;
+	padding-top: 1rem;
+	padding-bottom: 0;
+	text-align: center;
+	font-size: 1rem;
+  }
+  
+  .auth-form__message-error {
+	color: var(--tw-red-500);
+  }
+  
+  .auth-form__btn {
+	margin-top: 1rem;
+	padding: 0.75rem;
+	width: 100%;
+	background: var(--tw-indigo-600);
+  }
+  
+  .auth-form__message {
+	height: 2.25rem;
+	padding-top: 1rem;
+	padding-bottom: 0;
+	text-align: center;
+	font-size: 1rem;
+  }
+  
+  .auth-form__message-error {
+	color: var(--tw-red-500);
+  }
+  
+  .auth-form__btn {
+	width: 100%;
+  }
+  
+  .auth-form__type {
+	text-align: center;
+	margin-top: 2rem;
+  }
+  
+  .dashboard__wrapper {
+	width: 100vw;
+	height: 100vh;
+	display: flex;
+  }
+  
+  .dashboard__navigation {
+	width: 100%;
+	max-width: 300px;
+	height: 100%;
+	display: flex;
+	flex-direction: column;
+	justify-content: space-between;
+	color: var(--tw-slate-100);
+	background: linear-gradient(to bottom,
+		var(--tw-indigo-700),
+		var(--tw-blue-600),
+		var(--tw-indigo-700));
+	border-right: 1px solid var(--tw-slate-900);
+  }
+  
+  .dashboard__branding {
+	padding: 1rem;
+  }
+  
+  .dashboard__navigation-list,
+  .dashboard__navigation-admin-list {
+	padding: 1.5rem 1rem;
+  }
+  
+  .dashboard__branding {
+	font-size: 1.5rem;
+	text-align: center;
+	box-shadow: 0 25px 50px -12px rgb(0 0 0 / 0.25);
+  }
+  
+  .dashboard__navigation-list {
+	list-style: none;
+  }
+  
+  .dashboard__navigation-admin-separator {
+	border-top: 1px solid var(--tw-slate-100);
+	padding: 0 1rem;
+	width: 90%;
+	margin: 0 auto;
+	opacity: 0.5;
+  }
+  
+  .dashboard__navigation-admin-list {
+	list-style: none;
+  }
+  
+  .dashboard__navigation-item {
+	margin: 0.75rem 0;
+  }
+  
+  .dashboard__navigation-item:first-child {
+	margin-top: 0;
+	margin-bottom: 0.75rem;
+  }
+  
+  .dashboard__navigation-link {
+	color: var(--tw-slate-100);
+	background: none;
+	border: none;
+	font-size: 1.15rem;
+	cursor: pointer;
+	display: flex;
+	align-items: center;
+  }
+  
+  .dashboard__navigation-link>svg {
+	height: 1.5rem;
+	width: 1.5rem;
+	margin-right: 0.75rem;
+  }
+  
+  .dashboard__navigation-sign-out {
+	margin: 1rem;
+	padding: 0.5rem 1rem;
+	background: var(--tw-slate-100);
+	color: var(--tw-slate-900);
+	font-size: 1rem;
+  }
+  
+  .dashboard__content {
+	padding: 1.5rem 1rem;
+	width: 100%;
+	height: 100%;
+	background: var(--tw-slate-100);
+  }
+  
+  @media screen and (min-width: 768px) {
+  .nav__brand {
+	  font-size: 1.5rem;
+	}
+  
+	.hero__title {
+	  font-size: 4.25rem;
+	}
+  
+	.hero__intro {
+	  font-size: 1.5rem;
+	  text-align: center;
+	  width: 100%;
+	  max-width: 68ch;
+	}
+  
+	.waitlist-form {
+	  min-width: 32rem;
+	}
+  
+	.waitlist-form__group {
+	  flex-direction: row;
+	  display: flex;
+	  align-items: center;
+	}
+  
+	.waitlist-form__label {
+	  min-width: 24rem;
+	}
+  
+	.waitlist-form__input {
+	  margin-bottom: 0;
+	  width: 95%;
+	}
+  
+	.auth-form__title {
+	  font-size: 2.5rem;
+	}
+  }
